pyproject.toml
[tool.poetry]
name = "shift-task"
version = "0.1.0"
description = ""
authors = ["Ivan Ryabchenko"]
readme = "README.md"
packages = [{include = "shift_task"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.2"}
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.15"
pytest = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]